Update sysmem
split header and update function name
